requirements.txt
PyJWT==2.10.1
termcolor==3.1.0
